move StorageAddress to separate file
